Friday Fixes (#95)
* advanced fee mode

* rebuild

* Makes wording around the character limit more clear.

As a user, when I went to the site and found that my last name was unavailable because it was too short I was quite miffed.  I didn't find the wording to be clear enough that the limitations were temporary nor did I know where to go for more details.  This change just adds some emphasis, makes it painfully clear that the limit will eventually be reduced, and provides a link to more details.

* more checks that the dictionary is loaded

* adds note saying char limit is temp